Fix source priority regression
Rubygems sources also include locally installed specs, and that confuses
bundler when finding the right source for a dependency, and when
checking whether a resolved spec is present ambiguously in two different
sources.

I think we should probably split this "locally installed" fallback to a
separate fallback source, but that seems like a big refactoring so for
now I'm using the hacky approach of manually excluding locally installed
specs from rubygems sources when needed.